chore: RestStyledComponentを作成、metadataをlayoutに戻す
